Only schedule push-mirror updates once per push
Before we were scheduling the update once for each pushed ref through
the Git::(Branch|Tag)PushService. But all the refs will be pushed
every time, meaning that most of these updates would be no-ops.

This moves the updated out into the PostReceive worker, so this is
only scheduled once per push.