Add a link to WebIDE on project page in all cases
Add a link to the WebIDE on project page in case the user doesn't have
write access to the repo. Clicking the link will actually create a fork
first and then redirect to the WebIDE in the forked repo.

A confirm modal is shown before forking the project.

The button would not appear if the user cannot create a fork or an MR
in the fork.