Add table for grafana api tokens
In preparation for supporting the ability to embed Grafana metrics
in GitLab Flavored Markdown, we need to allow a way for users to
store credentials for us to access the HTTP API for their grafana
instance. We expect users to save two key attributes:
an authorization token and the url of their grafana instance.

This work adds the table and model for storing this auth config.