Optimize namespace.has_parent? method
We can avoid doing an extra SQL query for loading parent when
only checking if parent exists.

Because in Github import service we use nested setting of groups
(so there is a situation when `group` is set but `group_id` not),
we check presence of both group_id and group.