Add spdx_identifier to software_licenses
This adds the `spdx_identifier` column to the `software_licenses` table.
It also adds a sidekiq-cron to update our list of software licenses from
the SPDX database on a weekly basis. This list can offer our customers
a more comprehensive list of software licenses to choose from and
reduce the need for user entered software license names.