Add controller action for creating cluster AWS roles
Once hooked up to the Vue app, this will take care of
the first step of creating a cluster on EKS - capturing
the role GitLab will assume to create resources in the
user's account.

Adds four new options to ApplicationSetting:

* eks_integration_enabled
* eks_account_id
* eks_access_key_id
* eks_secret_access_key

:eks_account_id is exposed to the frontend to allow
users to add it as a trusted entity to their chosen
IAM role.