Remove gitlab-shell from test setup
Gitaly absorbed the hooks of shell. Before Gitaly would still depend on
the GitLab Shell config. However, this code is only executed when the
hooks are executed. During the tests we don't as that would require the
internal API to run.

These things combined means we could effectively remove the setup off
shell before tests. This makes the setup of tests faster and a big more
smooth.