Add praefect as a socket for tests
Praefect functions as a proxy, but there are a few RPCs that are only in
praefect and do not get proxied to gitaly. To test functionality that
hits these RPCs, we need a way to hit a praefect instance that talks to
a gitaly instance.

This MR sets up praefect in front of the existing gitaly that is used in
tests.