Refactor processing of ref changes in PostReceive
To support creating push events in bulk and aggregate them by
type (branch or tag) and action (created, pushed, deleted), we
need to be able to segregate the changes into groups.

Having that inside the `PostReceive` worker class will increase
its reponsibility further. To help with that, a new service
called `Git::ProcessRefChangesService` has been added.

It's now responsible for processing branch/tag changes and
delegate it to the appropriate push service class.

This will be later on used to determime whether we need to create
individual or bulk push events.