Fix erroneous "No activities found" message
If the number of activities on the page were less than the limit (20),
the pager would display the set of activities but also insert a "No
activities found" banner underneath. This was happening because the
pager attempted to load more data because it thought it needed to
display more information in the window.

This commit fixes the problem by disabling the pager if the number of
activities received is less than the total limit.

Closes https://gitlab.com/gitlab-org/gitlab/issues/27118