Allow adding requests to performance bar manually
Add a '+' button in the performance bar. When clicked, it allows you to
enter a request ID or a URL. A request ID will fetch performance bar
results for that request, and a URL will fetch that URL and then its
performance bar data.

Because of the way it's triggered, the URL will use the current user's
credentials, and it must be to a JSON endpoint - that's the content type
it will request. A request ID can be for any request (HTML, JSON, XML,
images) and can be for any user.