Remove test expectation due to feature change
Now issue boards hide redundant labels so the removed expectation
makes no sense anymore. Actually, due to this expectation this test
is failing on master.

This is not really related to what this MR intends to do, which is
to create a new e2e tests, but since I found it while creating this
new test, I thought it would make sense to fix it here.