Fetch user via API in fabricate_or_use
Previously fabricate_or_use would only create a User instance with the
provided username and password. This was fine when only those creds
were needed (e.g., to sign on), but it's not enough in tests that
need other user details, e.g., the user's name.

This changes fabricate_or_use so that it fetches the user via the API,
which provides access to the user's details.