Raise issuable text limits
The 16,000 and 48,000-byte description and description HTML limits were
a bit too low. In some of our onboarding tasks, the text took close to
170K.

Our API limits the description to be 1 megabyte, so update the limit to
match that. From
https://gitlab.com/gitlab-org/gitlab/merge_requests/16181, the estimated
description HTML multiple was 3x, so make this limit 5x to give some
leeway.