Redact search results based on Ability.allowed?
In order to improve security we want to avoid any chance of leaking
search results that the user should not have access to. This approach
should account for bugs from incorrect queries or stale/incorrect data
in the index.

It happens that this also fixes
https://gitlab.com/gitlab-org/gitlab/issues/33712 for the time being as
we have not yet fixed the root cause of the problem which is that the
query is wrong.