Fix Gitaly call duration measurements
In many cases, we were only measuring the time to return from the Gitaly
RPC, not the total execution time of the RPC for streaming responses. To
handle this, we wrap the RPC in a new method,
`GitalyClient.streaming_call` that yields a response and consumes it.

Only `CommitService` has been updated to use this new measurement.
Other services should be updated in subsequent merge requests.

Relates to https://gitlab.com/gitlab-org/gitlab/issues/30334