Update gitlab_rails_cheat_sheet.md
There is no example of the rails runner which demonstrates that you will need to use something like puts to actually print the personal_access_token.token 

Since the only example at the top is the runner, if you try the example generation of a personal access token. It will create the token, but not return the value. So you can't use it.