Add service to fetch and format a grafana embed
Adds a new Metrics Dashboard Service which fetches a grafana
dashboard and datasource based on an input url, reformatting
the output to be compatible with our existing dashboard
rendering behavior on the FE. Ultimately, this will be used
as a part of embedding prometheus metrics in GFM via grafana.

Note that this service introduces usage of ReactiveCaching in
the service as a result of external API calls. This will mean
that usage of this new service will need to be polled.