Add table and model for `zoom_meetings`
Add Active Record backed class `ZoomMeeting`s to persist and validate
zoom links associated to an issue. We also associate meetings to a
project (via project_id) to avoid table joins. `ZoomMeeting`s can have
an issue_status of `added` or `removed` which will be used to show or
hide the zoom link on an issue via quick actions.