Allow Gitaly to work without a PG connection
Gitaly didn't require a PG connection to connect to, though recently we
started to pass feature flags to Gitaly in different manners. That
change would now require a PG to be running. This is not ideal,
especially if only for feature flags.

This change rescues a bad connection error, and breaks the required
dependency between PG and Gitaly.

Closes https://gitlab.com/gitlab-org/gitlab/issues/34205