Add project option for deleting source branch
Feature branches should be deleted after being merged into a stable
branch like master. This is not universally true, and accidentally
deleting a branch cause frustration and confusion for users.

The change:
- adds a Project Setting to control if 'Delete source branch' option is
enabled or disabled for new merge requests
- the Project Setting is on for new and existing projects

The previous default behaviour can be restored by toggling the new
setting.

Historically, the behaviour was changed to delete the source branch as
part of the merge widget refactor and then reverted 3 months later.

- Widget refactor 8db76243abfd1ebc98553426075f3fb4ddab54fe
- Behaviour reverted fd72b4eda1f0d7d3591a6db8163633efaa3f1955