Disable protected path throttling by default
In GitLab 11.0 (see
https://gitlab.com/gitlab-org/gitlab-foss/issues/40973 and
https://gitlab.com/gitlab-org/omnibus-gitlab/merge_requests/2237), we
disabled Rack Attack by default in Omnibus.

We inadvertently enabled by default in GitLab 12.4 via
https://gitlab.com/gitlab-org/gitlab/merge_requests/16463. This commit
introduces a database migration to restore the default to off.

This this doesn't touch the existing settings, so admins will have to
know to switch it off if they upgraded to GitLab 12.4.

Closes https://gitlab.com/gitlab-org/gitlab/issues/34212