Update quarantined test so that it runs
While working on another MR I noticed that this quarantined test had a
bug that caused it to report the wrong number of arguments when
calling `within_submenu`.

This changes fixes the test, although it stays in quarantine because
the issue for the bug in the application is still open.

This change also updates the block args so the cop doesn't have to
be disabled.