Refactor search_tab? -> show_user_search_tab?
This method was only ever used with the argument :members but further it
seems to have been half-implemented generically and the argument wasn't
even being used so it was just causing confusion. Since this is only
meant to be checking if the users tab should show it is now a bit more
explicit.