Dynamically set db pool size in Puma
Before Puma, we had 1 rails process talk to the database
at any given time, which meant a connection pool size of 1
was sufficient to support that model. With Puma and its worker
thread pool, there can now be many threads simultaneously
trying to talk to the database. This change makes sure that
we allow the db pool size to scale along with the maximum
number of worker threads configured.