Remove empty Github service templates from db
When creating a new Github integration, it looks for a service
record of the same type, but with the template flag set to true,
and clone its attributes before presenting the record to the user.

`https://gitlab.com/gitlab-org/gitlab/merge_requests/15737/diffs`
added `static_context` as a default attribute on for
this service but it clashes with the empty template on .com.

This commit removes Github templates that have `nil` or `{}`
as properties, allowing new services to set the `static_context`
default value from the code.