Remove existence check from ProjectUrlConstrainer
This was causing non-existent projects to fall back to our
wildcard not found route which then behaves differently
for repository ref paths.

For repository ref paths like blobs, we ignore the format in the
URL because these are rendered as HTML. But when it falls back to the
not found route, the format is parsed causing Devise to use the
wrong status code for unauthenticated requests.

With this change, valid URLs whether the project exists or not will
be handled by the same controller so that behavior will be the same.

Only paths that have invalid project paths would fall back to
the wildcard catch-all route.