Wait for maven to finish deploying
The test builds a maven package in a docker container. The build
process can take a very different amount of time depending on how
long it takes to download all the dependencies and build.

There is an issue open to improve the test, but for now we make it
wait up to 5 minutes to build and deploy. We stop the container as
soon as deployment is complete so that it doesn't wait the full 5
minutes.

We also force delete the temp directory used to store the source for
the package because it seems to take some time for docker to release
the files after the container is stopped.