Make sure Projects#show is matched last
EE-routes should override CE ones so it should come first. This
also makes fallback routes in CE work efficiently.

For example: `/namespace/project/insights` matches
ProjectsController#show first and then fails due to
ProjectUrlConstrainer (when it was checking for existence)
then it tries to match the route defined in EE.

This results in unnecessary SQL queries when we had existence
checks and broken routes when we removed the existence check.

We also needed to move the definition of legacy routes so that
they are matched earlier.