Use WriteRef to create/delete branches w/o hooks
The GitLab::Git::Repository class had two methods which aren't really
used in production. `#create_branch` and `#remove_branch` now leverage
WriteRef. The behaviour does change now, in the case of `#create_branch`
it doesn't throw an error anymore if the branch already exists. That's
actually not a problem as all call sites are import related, in which
case there's only one branch to be imported.

`#delete_branch` is only used in tests. The runtime errors aren't
depended on in that case. Again, this change in behaviour doesn't
influence GitLab-Rails.