Do not cache uploads for six months by default
Setting a long-term cache header works fine when not using object
storage, or when using object storage with proxy_download enabled. When
using object storage _without_ proxy_download, we send a redirect that's
cached for six months to a signed URL in object storage, but the signed
URL only works for ten minutes. So once those ten minutes are up, the
browser will keep following the (cached) redirect to a dead link.

In future, we could be more sophisticated by setting the TTL based on
the settings we're using. For now, just set this back to the pre-12.4
state of not caching content for longer than 5 minutes.