Fix missing RSS feed events
The ProjectsController currently loads a list of projects the user has
access to in the function load_projects. It uses this list to render the
actual list of projects, but it also uses this list to render the
activity feed in the atom form for the RSS feed.

The problem is that load_projects already paginates the list of projects
which leads to wrong results (only the first 20 projects are considered)
when rendering the RSS feed.

This commit uses a different ProjectFinder for loading the events that
does not paginate the results and keeps the pagination in load_projects
because all the other function need paginated results.