Increase certmanager install timeout
We have been getting reports from users that the cluster
issuer is not ready prior to timing out, failing the
installation. Since this time is primarily spent on the
cluster and not in Sidekiq, we can probably afford to wait
a bit longer.

See e.g. https://gitlab.com/gitlab-org/gitlab/issues/33186