Nullify user roles with software developer value
On October 20th code was deployed to GitLab.com
which caused profile save actions from any user to
update the role attribute to a value of 0 (software
developer). From October 21st, 16:47 CEST until October
23rd, 19:11 CEST, the new signup flow experiment was
run, which forced 10% of users created in that
timeframe to pick a role. October 24th (merged October
28th, not yet deployed) a fix was created for the
unintentional saving of the role.

Solution: after the fix is deployed (check with /chatops
run auto_deploy status aac40d26), we should nullify all
oles with a value of 0. Other values are intentional and
should be kept. Unfortunately, we must also nullify roles
with a value of software developer for users created
within the timeframe the experiment was run, because,
based on the database, we cannot distinguish between
users that participated in the experiment and users in the
control group.