Update "view wiki pages" permission documentation
The "view wiki pages" permission for guests has always applied to
private projects, the documentation was previously incorrect.

The documentation was originally changed in 9b12562c however the
:read_wiki permissions at that same ref were the same as they are now
(allowing guests of private projects to read the wiki).

We also have a feature
https://gitlab.com/gitlab-org/gitlab/issues/18739
that explicitly serves the wiki to guests of private projects, so this
further prooves that this documentation was wrong.

https://gitlab.com/gitlab-org/gitlab/issues/34761