Update squash_commit_sha only on successful merge
With https://gitlab.com/gitlab-org/gitlab/merge_requests/17052 we
started saving and exposing squash commit sha for squash & fast-forward
merge requests but we do it even if the merge fails. This commit fixes
this and add tests to ensure we do update the column only if the merge
was successful.

Related to https://gitlab.com/gitlab-org/gitlab/issues/35709.