Authenticate Runtime::Feature as admin
Uses an admin access token if one is provided (via
GITLAB_QA_ADMIN_ACCESS_TOKEN)

Otherwise it uses the admin username and password via Runtime::Env.
Those credentials fallback on the default username and password if
GITLAB_ADMIN_USERNAME and GITLAB_ADMIN_PASSWORD are not set, so it
checks if the user with those credentials has admin access, and raises
an exception if it doesn't.