Remove ES parent relation for project snippets
https://gitlab.com/gitlab-org/gitlab/merge_requests/18459 added a join
relation on project snippets to support searching public/internal
projects.

Since this requires recreating the index this broke indexing on existing
indices, so we decided to revert this change for now.

We're keeping most of the changes from the original MR since it also
refactored the ES snippet classes to align with the other models.