Allow DesignResolver to filter designs by filename
This allows a client to filter designs by their filenames.

https://gitlab.com/gitlab-org/gitlab/merge_requests/18997 added the
ability to filter designs by their (global) IDs, however in order to
not break existing routes to designs in GitLab that are currently using
the design's filename in the route rather than their IDs, we also need
the resolver to be able to filter designs by filename.

https://gitlab.com/gitlab-org/gitlab/issues/34408