Fix Bitbucket Cloud importer pull request state
Due to e404c31152cb6ce7acc82969ba103505912946ce, all pull requests are
created in the `opened` state because assigning the `state` parameter
when creating a `MergeRequest` no longer works. To fix this, define the
`state_id` in the `initalize` step to ensure state_machine has the
right value to use.

Closes https://gitlab.com/gitlab-org/gitlab/issues/35746