Introduce project_settings table
There is no usable data in this table yet, but with this change the
table is created, it's backfilled for every existing project, and new
rows are created as new projects are created. When the background
migration is complete it should be fairly easy to add columns to this
table and store project settings data in this table instead of in the
projects table itself.

This is a first step toward:
https://gitlab.com/gitlab-org/gitlab/issues/19826