Use CI_DEFAULT_BRANCH in Auto DevOps templates
We now have a CI_DEFAULT_BRANCH variable for use in our Auto DevOps
CI templates, which allows us to code logic for any default branch,
not just those named `master`.

This change was previously released with the intention that the
variable would be interpolated under the `ref` key of `only/except`,
but that didn't work so I reverted it. Now it exists under the
`variables` key, where it seems to be working.

https://gitlab.com/gitlab-org/gitlab/issues/33601