Add login flow
Adds a Flow module for frequently used sequences of actions

Adds a login flow, including for logging in, performing some actions
and then logging out.

Updates a test as an example of how to use the login flow.

Adds a class method `QA::Resource::User.admin` to provide access
to the admin user's details (include username/password needed to
sign in)