Fix CSS loading errors in e-mail notifications
https://gitlab.com/gitlab-org/gitlab/merge_requests/17699 introduced a
new CSS file that was not rendered properly in production mode since
`Rails.configuration.assets.compile` is disabled and the search path was
not correct.

To fix this, we use `Rails.root.join` instead of `File.join`. The latter
provides a relative path that is not found by the Rails search path.

Closes https://gitlab.com/gitlab-org/gitlab/issues/36201