Use explicitly tracked deployments for MR widgets
When displaying the deployments of a merge request in the MR widget, we
used to rely on the latest CI pipeline to determine what the deployment
was. In this commit we add the ability to use the table
"deployment_merge_requests" to display the deployments of a merge
request. This way the data displayed remains the same, regardless of the
state of any CI pipelines.

These changes are currently hidden behind the
"deployment_merge_requests_widget" project feature flag. We do not
reuse the existing deployments feature flag, as we still want to record
merge requests even if the UI is not able to display the data properly.