Migrate HighlightCache -> DeprecatedHighlightCache
Originally I had hoped to feature-flag the changes to support redis hash
caching, as they seemed small, but of course those changes expanded and
expanded until it became clear it really needed to be extracted into a
new class with a parallel API to keep it clean. h/t to @oswaldo for
pointing out when we'd crossed that threshold!