Remove WorkingCopyBaseService
The working copy for squash and rebases, among other git operations, has
been a concern of Gitaly for a long time. This service is an old
artifact of the time GitLab-Rails was responsible for it.

This service contains little code, and the code it did contain I've
moved to the usual base service.

Closes: https://gitlab.com/gitlab-org/gitlab/issues/36067