Use commit SHA instead of branch id
Use the commit SHA instead of the branch ID to request data for
the file tree and raw files in the IDE.  The SHA is obtained as
the most recent commit in the branch at the moment the branch is
first loaded, then remains the same until the branch is changed
or the window is refreshed.  This is done to prevent unusual file
behavior that could occur if another user edits the files while
the IDE is open.

NOTE:

We opted to build the raw path on demand rather than keep it in
the file object because when a new commit is made, it'd take a
lot of processing to update all the paths.

For this reason, this MR adds a few long `joinPaths` calls
to actions like `getFiles`.