Remove environment_logs_use_vue_ui feature flag
The Pods logs page was refactored to VueJS for the "Pod Logs" page.
Previously, the Logs page was done using HAML and jQuery.

The rewrite work was done behind feature flag
environment_logs_use_vue_ui.

This commit is the cleanup step in which the feature
flag is dropped, and the VueJS counterpart is set as the
default for users, so old unused code is removed:

- Remove old haml and jquery based feature files
- Update spec to use new js-* classes
- Extend feature spec to include the environments dropdown
- Remove unused scrolling class
- Remove unused static fixture
- Remove unused locale strings